CustomSprite分離 _vsnprintfを_vsnprintf_sに置き換え
