+++ samples/DXLibrary/basic/SSPlayer/common/Animator/ssplayer_effect.cpp
@@ -9,6 +9,7 @@
 #include "ssplayer_effectfunction.h"
 
 #include "../../SS5PlayerPlatform.h"
+#include "common/SS5PlayerLibs/SS5CustomSprite.h"
 
 
 namespace ss
